refactor: move clock/clock.rs into clock.rs (clippy)
