--- src/protocol/clock/mod.rs
@@ -1,1 +1,0 @@
-pub mod clock;
